Update makefile, add simple go program :triangular_flag_on_post:
Started using the ways of literate programming by harnessing the power of org
but it is not a good idea for real development since you don't get to have the
major mode benefits of the actual language that you are writing the code for.
You may be able to get it rolling with few configurations but since I am a
beginner in emacs, I will have to resort to write comments in the .go programms
instead of writing go programms in .org files.
Makefiles are awesome!
I am following along the tutorials from ~sentdex~ and I hope to get it all
covered in a week or so, unless I run into some challenges allocating time to
learn this new language.
Org mode is awesome! I have only scratched the surface but I am sure I will dig
deeper once I bring it up in my day-to-day workflow.