finish type related quest :white_check_mark:
I found go to be similar to typescript but I guess since I haven't used a strict
typing language in a while, any statically typed language will look similar to
go right now. I love the way we can compress/club similar types so that we don't
have to write them again and again. A feature that I am excited to used is the
multiple returns. One strict thing that I found was the unused variables errors,
In C language, If I remember correctly, we had compile type warnings saying that
there were some unused variables in our program but it didn't stop the
compilation, but in go, on the other hand, it throws an error for unused
variables, I am not sure if we can disable that by using some compile flag but I
guess I am not too concerned about that right now.